refactor: move prune orchestration to snapshots/cmd_handler/ package
