+++ nbkp/snapshots/cli/prune_cmd.py
@@ -12,7 +12,7 @@ from ...clihelpers import OutputFormat, load_config_or_exit
 from ...clihelpers.endpoints import resolve_endpoints
 from ...config.epresolution import NetworkType
 from ...sync.output import print_human_prune_results
-from ..pruner import prune_all_syncs
+from ..cmd_handler import prune_all_syncs
 
 # Transitional imports — these will move out of cli.common later.
 from ...cli.common import check_all_with_progress, managed_mount
