+++ nbkp/snapshots/cmd_handler/__init__.py
@@ -1,0 +1,3 @@
+"""Snapshot command handlers: orchestration logic for CLI commands."""
+
+from .prune import prune_all_syncs as prune_all_syncs
