+++ nbkp/snapshots/cmd_handler/prune.py
@@ -2,13 +2,13 @@
 
 from __future__ import annotations
 
-from ..config import Config
-from ..config.epresolution import ResolvedEndpoints
-from ..preflight import SyncStatus
-from .models import PruneResult
-from .btrfs import prune_snapshots as btrfs_prune_snapshots
-from .common import list_snapshots
-from .hardlinks import prune_snapshots as hl_prune_snapshots
+from ...config import Config
+from ...config.epresolution import ResolvedEndpoints
+from ...preflight import SyncStatus
+from ..models import PruneResult
+from ..btrfs import prune_snapshots as btrfs_prune_snapshots
+from ..common import list_snapshots
+from ..hardlinks import prune_snapshots as hl_prune_snapshots
 
 
 def _skip_reason(
