+++ tests/snapshots/test_cli.py
@@ -81,8 +81,8 @@ def _prune_active_statuses(
 
 
 class TestPruneCommand:
-    @patch("nbkp.snapshots.pruner.list_snapshots")
-    @patch("nbkp.snapshots.pruner.btrfs_prune_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.list_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.btrfs_prune_snapshots")
     @patch("nbkp.cli.common.check_all_syncs")
     @patch("nbkp.clihelpers.config.load_config")
     def test_successful_prune(
@@ -109,8 +109,8 @@ class TestPruneCommand:
         assert "OK" in result.output
         mock_prune.assert_called_once()
 
-    @patch("nbkp.snapshots.pruner.list_snapshots")
-    @patch("nbkp.snapshots.pruner.btrfs_prune_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.list_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.btrfs_prune_snapshots")
     @patch("nbkp.cli.common.check_all_syncs")
     @patch("nbkp.clihelpers.config.load_config")
     def test_dry_run(
@@ -142,8 +142,8 @@ class TestPruneCommand:
         call_kwargs = mock_prune.call_args
         assert call_kwargs.kwargs.get("dry_run") is True
 
-    @patch("nbkp.snapshots.pruner.list_snapshots")
-    @patch("nbkp.snapshots.pruner.btrfs_prune_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.list_snapshots")
+    @patch("nbkp.snapshots.cmd_handler.prune.btrfs_prune_snapshots")
     @patch("nbkp.cli.common.check_all_syncs")
     @patch("nbkp.clihelpers.config.load_config")
     def test_json_output(
